speech-whisper/requirements.txt
fastapi
pyyaml
aiohttp
python-multipart
openai==0.28.1
uvicorn
